requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.0

# Agent & extraction
docling>=2.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain>=0.2.0

# Document generation
reportlab>=4.0.0
python-docx>=1.1.0

# Config & logging
pyyaml>=6.0.0

# Production
gunicorn>=21.0.0
